[IMP]: add sererate ActivityListItem and refactor List component to display activities grouped by date add field description to Model

+++ client-app/src/features/activities/dashboard/ActivityList.tsx
@@ -1,55 +1,32 @@
-import { Button, Item, Label, Segment } from "semantic-ui-react";
-import { IActivity } from "../../../app/models/Activity";
+import { observer } from "mobx-react-lite";
+import { Header, Item, Segment } from "semantic-ui-react";
 import { useStore } from "../../../app/stores/Store";
 import { ConfirmModal } from "../form/ConfirmModal";
-import { useState } from "react";
-import { observer } from "mobx-react-lite";
-import { Link } from "react-router-dom";
+import { ActivityListItem } from "./ActivityListItem";
 
-export const ActivityList = observer(function() {
+export const ActivityList = observer(function () {
   const { activityStore } = useStore();
-  const [todeleteActivity, setTodeleteActivity] = useState<IActivity | undefined>(undefined);
+  const { todeleteActivity, setTodeleteActivity } = activityStore;
 
   return (
-    <Segment>
-      <Item.Group divided>
-        {[...activityStore.activities].map(([activityId, activity]) => (
-          <Item key={activityId}>
-            <Item.Content>
-              <Item.Header as="a">{activity.title}</Item.Header>
-              <Item.Meta>{activity.date.format("LLL")}</Item.Meta>
-              <Item.Description>
-                {activity.city}, {activity.venue}
-              </Item.Description>
-              <Item.Extra>
-                <Button
-                  floated="right"
-                  content="view"
-                  color="blue"
-                  as={Link} to={`${activity.id}`}
-                ></Button>
-                <Button
-                  floated="right"
-                  content="remove"
-                  color="red"
-                  onClick={() => setTodeleteActivity(activity)}
-                ></Button>
-                <Label basic content={activity.category}></Label>
-              </Item.Extra>
-            </Item.Content>
-          </Item>
-        ))}
-      </Item.Group>
+    <>
+      {[...activityStore.activitiesGroupedBy("date")].map(([date, activities]) => (
+        <>
+          <Header sub color="teal">{date}</Header>
+            {activities.map((activity) => (
+              <ActivityListItem key={activity.id} activity={activity}></ActivityListItem>
+            ))}
+        </>
+      ))}
       <ConfirmModal
         title={"Deleting Activity"}
         onClickYes={() => {
-          if(todeleteActivity) 
-            activityStore.deleteActivity(todeleteActivity); 
+          if (todeleteActivity) activityStore.deleteActivity(todeleteActivity);
           setTodeleteActivity(undefined);
         }}
         onClickNo={() => setTodeleteActivity(undefined)}
         isOpen={todeleteActivity != undefined}
       ></ConfirmModal>
-    </Segment>
+    </>
   );
 });
